link: don't clobber existing files

+++ internal/appherder/link.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
+	"strings"
 )
 
 func (a App) linkPath(appName string) string {
@@ -24,9 +25,19 @@ func (a App) Link(appName string) error {
 		return fmt.Errorf("create %s: %w", a.binDir, err)
 	}
 	dst := a.linkPath(appName)
-	if err := os.Remove(dst); err != nil && !os.IsNotExist(err) {
-		return fmt.Errorf("remove existing link %s: %w", dst, err)
+
+	if target, err := os.Readlink(dst); err == nil {
+		if target == src {
+			return nil
+		}
+		fmt.Fprintf(os.Stderr, "Warning: %s is already linked to a different target\n", dst)
+		return nil
 	}
+	if _, err := os.Lstat(dst); err == nil {
+		fmt.Fprintf(os.Stderr, "Warning: %s already exists, skipping\n", dst)
+		return nil
+	}
+
 	if err := os.Symlink(src, dst); err != nil {
 		return fmt.Errorf("link %s -> %s: %w", dst, src, err)
 	}
@@ -36,6 +47,15 @@ func (a App) Link(appName string) error {
 func (a App) Unlink(appName string) error {
 	appName = NormalizeAppName(appName)
 	dst := a.linkPath(appName)
+
+	target, err := os.Readlink(dst)
+	if err != nil {
+		return fmt.Errorf("%s is not a link managed by appherder", dst)
+	}
+	if !strings.HasPrefix(target, a.appimagesDir+string(filepath.Separator)) {
+		return fmt.Errorf("%s is not a link managed by appherder", dst)
+	}
+
 	if err := os.Remove(dst); err != nil {
 		return fmt.Errorf("remove link %s: %w", dst, err)
 	}
